Split SyncFn and ReSync tests
