+++ rest/adminapitest/admin_api_test.go
@@ -18,7 +18,6 @@ import (
 	"strconv"
 	"strings"
 	"sync"
-	"sync/atomic"
 	"testing"
 	"time"
 
@@ -505,532 +504,6 @@ func TestDBGetConfigNames(t *testing.T) {
 		assert.Equal(t, k, *v.Name)
 	}
 
-}
-
-// Take DB offline and ensure can post _resync
-func TestDBOfflinePostResync(t *testing.T) {
-
-	rt := rest.NewRestTester(t, nil)
-	defer rt.Close()
-
-	log.Printf("Taking DB offline")
-	response := rt.SendAdminRequest("GET", "/db/", "")
-	var body db.Body
-	require.NoError(t, base.JSONUnmarshal(response.Body.Bytes(), &body))
-	assert.True(t, body["state"].(string) == "Online")
-
-	response = rt.SendAdminRequest("POST", "/db/_offline", "")
-	rest.RequireStatus(t, response, 200)
-
-	response = rt.SendAdminRequest("GET", "/db/", "")
-	body = nil
-	require.NoError(t, base.JSONUnmarshal(response.Body.Bytes(), &body))
-	assert.True(t, body["state"].(string) == "Offline")
-
-	rest.RequireStatus(t, rt.SendAdminRequest("POST", "/db/_resync?action=start", ""), 200)
-	err := rt.WaitForCondition(func() bool {
-		response := rt.SendAdminRequest("GET", "/db/_resync", "")
-		var status db.ResyncManagerResponse
-		err := json.Unmarshal(response.BodyBytes(), &status)
-		assert.NoError(t, err)
-
-		var val interface{}
-		_, err = rt.Bucket().Get(rt.GetDatabase().ResyncManager.GetHeartbeatDocID(t), &val)
-
-		return status.State == db.BackgroundProcessStateCompleted && base.IsDocNotFoundError(err)
-	})
-	assert.NoError(t, err)
-}
-
-// Take DB offline and ensure only one _resync can be in progress
-func TestDBOfflineSingleResync(t *testing.T) {
-
-	syncFn := `
-	function(doc) {
-		channel("x")
-	}`
-	rt := rest.NewRestTester(t, &rest.RestTesterConfig{SyncFn: syncFn})
-	defer rt.Close()
-
-	// create documents in DB to cause resync to take a few seconds
-	for i := 0; i < 1000; i++ {
-		rt.CreateDoc(t, fmt.Sprintf("doc%v", i))
-	}
-	assert.Equal(t, int64(1000), rt.GetDatabase().DbStats.Database().SyncFunctionCount.Value())
-
-	log.Printf("Taking DB offline")
-	response := rt.SendAdminRequest("GET", "/db/", "")
-	var body db.Body
-	require.NoError(t, base.JSONUnmarshal(response.Body.Bytes(), &body))
-	assert.True(t, body["state"].(string) == "Online")
-
-	response = rt.SendAdminRequest("POST", "/db/_offline", "")
-	rest.RequireStatus(t, response, 200)
-
-	response = rt.SendAdminRequest("GET", "/db/", "")
-	body = nil
-	require.NoError(t, base.JSONUnmarshal(response.Body.Bytes(), &body))
-	assert.True(t, body["state"].(string) == "Offline")
-
-	response = rt.SendAdminRequest("POST", "/db/_resync?action=start", "")
-	rest.RequireStatus(t, response, http.StatusOK)
-
-	// Send a second _resync request.  This must return a 400 since the first one is blocked processing
-	rest.RequireStatus(t, rt.SendAdminRequest("POST", "/db/_resync?action=start", ""), 503)
-
-	err := rt.WaitForCondition(func() bool {
-		response := rt.SendAdminRequest("GET", "/db/_resync", "")
-		var status db.ResyncManagerResponse
-		err := json.Unmarshal(response.BodyBytes(), &status)
-		assert.NoError(t, err)
-
-		var val interface{}
-		_, err = rt.Bucket().Get(rt.GetDatabase().ResyncManager.GetHeartbeatDocID(t), &val)
-
-		return status.State == db.BackgroundProcessStateCompleted && base.IsDocNotFoundError(err)
-	})
-	assert.NoError(t, err)
-
-	assert.Equal(t, int64(2000), rt.GetDatabase().DbStats.Database().SyncFunctionCount.Value())
-}
-
-func TestResync(t *testing.T) {
-	base.LongRunningTest(t)
-
-	testCases := []struct {
-		name               string
-		docsCreated        int
-		expectedSyncFnRuns int
-		expectedQueryCount int
-		queryLimit         int
-	}{
-		{
-			name:               "Docs 0, Limit Default",
-			docsCreated:        0,
-			expectedSyncFnRuns: 0,
-			expectedQueryCount: 1,
-			queryLimit:         db.DefaultQueryPaginationLimit,
-		},
-		{
-			name:               "Docs 1000, Limit Default",
-			docsCreated:        1000,
-			expectedSyncFnRuns: 2000,
-			expectedQueryCount: 1,
-			queryLimit:         db.DefaultQueryPaginationLimit,
-		},
-		{
-			name:               "Docs 1000, Limit 10",
-			docsCreated:        1000,
-			expectedSyncFnRuns: 2000,
-			expectedQueryCount: 101,
-			queryLimit:         10,
-		},
-	}
-
-	syncFn := `
-	function(doc) {
-		channel("x")
-	}`
-
-	for _, testCase := range testCases {
-		t.Run(testCase.name, func(t *testing.T) {
-			rt := rest.NewRestTester(t,
-				&rest.RestTesterConfig{
-					DatabaseConfig: &rest.DatabaseConfig{DbConfig: rest.DbConfig{
-						QueryPaginationLimit: &testCase.queryLimit,
-					}},
-					SyncFn: syncFn,
-				},
-			)
-			defer rt.Close()
-
-			for i := 0; i < testCase.docsCreated; i++ {
-				rt.CreateDoc(t, fmt.Sprintf("doc%d", i))
-			}
-
-			response := rt.SendAdminRequest("POST", "/db/_resync?action=start", "")
-			rest.RequireStatus(t, response, http.StatusServiceUnavailable)
-
-			response = rt.SendAdminRequest("POST", "/db/_offline", "")
-			rest.RequireStatus(t, response, http.StatusOK)
-
-			rest.WaitAndAssertCondition(t, func() bool {
-				state := atomic.LoadUint32(&rt.GetDatabase().State)
-				return state == db.DBOffline
-			})
-
-			response = rt.SendAdminRequest("POST", "/db/_resync?action=start", "")
-			rest.RequireStatus(t, response, http.StatusOK)
-
-			var resyncManagerStatus db.ResyncManagerResponse
-			err := rt.WaitForCondition(func() bool {
-				response := rt.SendAdminRequest("GET", "/db/_resync", "")
-				err := json.Unmarshal(response.BodyBytes(), &resyncManagerStatus)
-				assert.NoError(t, err)
-
-				var val interface{}
-				_, err = rt.Bucket().Get(rt.GetDatabase().ResyncManager.GetHeartbeatDocID(t), &val)
-
-				if resyncManagerStatus.State == db.BackgroundProcessStateCompleted && base.IsDocNotFoundError(err) {
-					return true
-				} else {
-					t.Logf("resyncManagerStatus.State != %v: %v - err:%v", db.BackgroundProcessStateCompleted, resyncManagerStatus.State, err)
-					return false
-				}
-			})
-			assert.NoError(t, err)
-
-			assert.Equal(t, testCase.expectedSyncFnRuns, int(rt.GetDatabase().DbStats.Database().SyncFunctionCount.Value()))
-
-			var queryName string
-			if base.TestsDisableGSI() {
-				queryName = fmt.Sprintf(base.StatViewFormat, db.DesignDocSyncGateway(), db.ViewChannels)
-			} else {
-				queryName = db.QueryTypeChannels
-			}
-
-			assert.Equal(t, testCase.expectedQueryCount, int(rt.GetDatabase().DbStats.Query(queryName).QueryCount.Value()))
-			assert.Equal(t, testCase.docsCreated, resyncManagerStatus.DocsProcessed)
-			assert.Equal(t, 0, resyncManagerStatus.DocsChanged)
-		})
-	}
-
-}
-
-func TestResyncErrorScenarios(t *testing.T) {
-
-	if !base.UnitTestUrlIsWalrus() {
-		// Limitation of setting LeakyBucket on RestTester
-		t.Skip("This test only works with walrus")
-	}
-
-	syncFn := `
-	function(doc) {
-		channel("x")
-	}`
-
-	leakyTestBucket := base.GetTestBucket(t).LeakyBucketClone(base.LeakyBucketConfig{})
-
-	rt := rest.NewRestTester(t,
-		&rest.RestTesterConfig{
-			SyncFn:           syncFn,
-			CustomTestBucket: leakyTestBucket,
-		},
-	)
-	defer rt.Close()
-
-	leakyBucket, ok := base.AsLeakyBucket(rt.Bucket())
-	require.Truef(t, ok, "Wanted *base.LeakyBucket but got %T", leakyTestBucket.Bucket)
-
-	var (
-		useCallback   bool
-		callbackFired bool
-	)
-
-	if base.TestsDisableGSI() {
-		leakyBucket.SetPostQueryCallback(func(ddoc, viewName string, params map[string]interface{}) {
-			if useCallback {
-				callbackFired = true
-				response := rt.SendAdminRequest("POST", "/db/_resync?action=start", "")
-				rest.RequireStatus(t, response, http.StatusServiceUnavailable)
-				useCallback = false
-			}
-		})
-	} else {
-		leakyBucket.SetPostN1QLQueryCallback(func() {
-			if useCallback {
-				callbackFired = true
-				response := rt.SendAdminRequest("POST", "/db/_resync?action=start", "")
-				rest.RequireStatus(t, response, http.StatusServiceUnavailable)
-				useCallback = false
-			}
-		})
-	}
-
-	for i := 0; i < 1000; i++ {
-		rt.CreateDoc(t, fmt.Sprintf("doc%d", i))
-	}
-
-	response := rt.SendAdminRequest("GET", "/db/_resync", "")
-	rest.RequireStatus(t, response, http.StatusOK)
-
-	response = rt.SendAdminRequest("POST", "/db/_resync?action=start", "")
-	rest.RequireStatus(t, response, http.StatusServiceUnavailable)
-
-	response = rt.SendAdminRequest("POST", "/db/_resync?action=stop", "")
-	rest.RequireStatus(t, response, http.StatusBadRequest)
-
-	response = rt.SendAdminRequest("POST", "/db/_offline", "")
-	rest.RequireStatus(t, response, http.StatusOK)
-
-	rest.WaitAndAssertCondition(t, func() bool {
-		state := atomic.LoadUint32(&rt.GetDatabase().State)
-		return state == db.DBOffline
-	})
-
-	useCallback = true
-	response = rt.SendAdminRequest("POST", "/db/_resync?action=start", "")
-	rest.RequireStatus(t, response, http.StatusOK)
-
-	rest.WaitAndAssertBackgroundManagerState(t, db.BackgroundProcessStateCompleted, rt.GetDatabase().ResyncManager.GetRunState)
-	rest.WaitAndAssertBackgroundManagerExpiredHeartbeat(t, rt.GetDatabase().ResyncManager)
-
-	response = rt.SendAdminRequest("POST", "/db/_resync?action=stop", "")
-	rest.RequireStatus(t, response, http.StatusBadRequest)
-
-	response = rt.SendAdminRequest("POST", "/db/_resync?action=invalid", "")
-	rest.RequireStatus(t, response, http.StatusBadRequest)
-
-	// Test empty action, should default to start
-	response = rt.SendAdminRequest("POST", "/db/_resync", "")
-	rest.RequireStatus(t, response, http.StatusOK)
-
-	rest.WaitAndAssertBackgroundManagerState(t, db.BackgroundProcessStateCompleted, rt.GetDatabase().ResyncManager.GetRunState)
-	rest.WaitAndAssertBackgroundManagerExpiredHeartbeat(t, rt.GetDatabase().ResyncManager)
-
-	assert.True(t, callbackFired, "expecting callback to be fired")
-}
-
-func TestResyncStop(t *testing.T) {
-
-	if !base.UnitTestUrlIsWalrus() {
-		// Limitation of setting LeakyBucket on RestTester
-		t.Skip("This test only works with walrus")
-	}
-
-	syncFn := `
-	function(doc) {
-		channel("x")
-	}`
-
-	leakyTestBucket := base.GetTestBucket(t).LeakyBucketClone(base.LeakyBucketConfig{})
-
-	rt := rest.NewRestTester(t,
-		&rest.RestTesterConfig{
-			SyncFn: syncFn,
-			DatabaseConfig: &rest.DatabaseConfig{DbConfig: rest.DbConfig{
-				QueryPaginationLimit: base.IntPtr(10),
-			}},
-			CustomTestBucket: leakyTestBucket,
-		},
-	)
-	defer rt.Close()
-
-	leakyBucket, ok := base.AsLeakyBucket(rt.Bucket())
-	require.Truef(t, ok, "Wanted *base.LeakyBucket but got %T", leakyTestBucket.Bucket)
-
-	var (
-		useCallback   bool
-		callbackFired bool
-	)
-
-	if base.TestsDisableGSI() {
-		leakyBucket.SetPostQueryCallback(func(ddoc, viewName string, params map[string]interface{}) {
-			if useCallback {
-				callbackFired = true
-				response := rt.SendAdminRequest("POST", "/db/_resync?action=stop", "")
-				rest.RequireStatus(t, response, http.StatusOK)
-				useCallback = false
-			}
-		})
-	} else {
-		leakyBucket.SetPostN1QLQueryCallback(func() {
-			if useCallback {
-				callbackFired = true
-				response := rt.SendAdminRequest("POST", "/db/_resync?action=stop", "")
-				rest.RequireStatus(t, response, http.StatusOK)
-				useCallback = false
-			}
-		})
-	}
-
-	for i := 0; i < 1000; i++ {
-		rt.CreateDoc(t, fmt.Sprintf("doc%d", i))
-	}
-
-	err := rt.WaitForCondition(func() bool {
-		return int(rt.GetDatabase().DbStats.Database().SyncFunctionCount.Value()) == 1000
-	})
-	assert.NoError(t, err)
-
-	response := rt.SendAdminRequest("POST", "/db/_offline", "")
-	rest.RequireStatus(t, response, http.StatusOK)
-
-	rest.WaitAndAssertCondition(t, func() bool {
-		state := atomic.LoadUint32(&rt.GetDatabase().State)
-		return state == db.DBOffline
-	})
-
-	useCallback = true
-	response = rt.SendAdminRequest("POST", "/db/_resync?action=start", "")
-	rest.RequireStatus(t, response, http.StatusOK)
-
-	rest.WaitAndAssertBackgroundManagerState(t, db.BackgroundProcessStateStopped, rt.GetDatabase().ResyncManager.GetRunState)
-	rest.WaitAndAssertBackgroundManagerExpiredHeartbeat(t, rt.GetDatabase().ResyncManager)
-
-	assert.True(t, callbackFired, "expecting callback to be fired")
-
-	syncFnCount := int(rt.GetDatabase().DbStats.Database().SyncFunctionCount.Value())
-	assert.True(t, syncFnCount < 2000, "Expected syncFnCount < 2000 but syncFnCount=%d", syncFnCount)
-}
-
-func TestResyncRegenerateSequences(t *testing.T) {
-
-	base.LongRunningTest(t)
-	syncFn := `
-	function(doc) {
-		if (doc.userdoc){
-			channel("channel_1")
-		}
-	}`
-
-	base.SetUpTestLogging(t, base.LevelInfo, base.KeyAll)
-
-	var testBucket *base.TestBucket
-
-	if base.UnitTestUrlIsWalrus() {
-		var closeFn func()
-		testBucket, closeFn = base.GetPersistentWalrusBucket(t)
-		defer closeFn()
-	} else {
-		testBucket = base.GetTestBucket(t)
-	}
-
-	rt := rest.NewRestTester(t,
-		&rest.RestTesterConfig{
-			SyncFn:           syncFn,
-			CustomTestBucket: testBucket,
-		},
-	)
-	defer rt.Close()
-
-	var response *rest.TestResponse
-	var docSeqArr []float64
-	var body db.Body
-
-	for i := 0; i < 10; i++ {
-		docID := fmt.Sprintf("doc%d", i)
-		rt.CreateDoc(t, docID)
-
-		response = rt.SendAdminRequest("GET", "/db/_raw/"+docID, "")
-		require.Equal(t, http.StatusOK, response.Code)
-
-		err := json.Unmarshal(response.BodyBytes(), &body)
-		require.NoError(t, err)
-
-		docSeqArr = append(docSeqArr, body["_sync"].(map[string]interface{})["sequence"].(float64))
-	}
-
-	role := "role1"
-	response = rt.SendAdminRequest("PUT", fmt.Sprintf("/db/_role/%s", role), fmt.Sprintf(`{"name":"%s", "admin_channels":["channel_1"]}`, role))
-	rest.RequireStatus(t, response, http.StatusCreated)
-
-	username := "user1"
-	response = rt.SendAdminRequest("PUT", fmt.Sprintf("/db/_user/%s", username), fmt.Sprintf(`{"name":"%s", "password":"letmein", "admin_channels":["channel_1"], "admin_roles": ["%s"]}`, username, role))
-	rest.RequireStatus(t, response, http.StatusCreated)
-
-	_, err := rt.Bucket().Get(base.RolePrefix+"role1", &body)
-	assert.NoError(t, err)
-	role1SeqBefore := body["sequence"].(float64)
-
-	_, err = rt.Bucket().Get(base.UserPrefix+"user1", &body)
-	assert.NoError(t, err)
-	user1SeqBefore := body["sequence"].(float64)
-
-	response = rt.SendAdminRequest("PUT", "/db/userdoc", `{"userdoc": true}`)
-	rest.RequireStatus(t, response, http.StatusCreated)
-
-	response = rt.SendAdminRequest("PUT", "/db/userdoc2", `{"userdoc": true}`)
-	rest.RequireStatus(t, response, http.StatusCreated)
-
-	// Let everything catch up before opening changes feed
-	require.NoError(t, rt.WaitForPendingChanges())
-
-	type ChangesResp struct {
-		Results []struct {
-			ID  string `json:"id"`
-			Seq int    `json:"seq"`
-		} `json:"results"`
-		LastSeq string `json:"last_seq"`
-	}
-
-	changesRespContains := func(changesResp ChangesResp, docid string) bool {
-		for _, resp := range changesResp.Results {
-			if resp.ID == docid {
-				return true
-			}
-		}
-		return false
-	}
-
-	var changesResp ChangesResp
-	request, _ := http.NewRequest("GET", "/db/_changes", nil)
-	request.SetBasicAuth("user1", "letmein")
-	response = rt.Send(request)
-	rest.RequireStatus(t, response, http.StatusOK)
-	err = json.Unmarshal(response.BodyBytes(), &changesResp)
-	assert.Len(t, changesResp.Results, 3)
-	assert.True(t, changesRespContains(changesResp, "userdoc"))
-	assert.True(t, changesRespContains(changesResp, "userdoc2"))
-
-	response = rt.SendAdminRequest("GET", "/db/_resync", "")
-	rest.RequireStatus(t, response, http.StatusOK)
-
-	response = rt.SendAdminRequest("POST", "/db/_offline", "")
-	rest.RequireStatus(t, response, http.StatusOK)
-
-	response = rt.SendAdminRequest("POST", "/db/_resync?action=start&regenerate_sequences=true", "")
-	rest.RequireStatus(t, response, http.StatusOK)
-
-	rest.WaitAndAssertBackgroundManagerState(t, db.BackgroundProcessStateCompleted, rt.GetDatabase().ResyncManager.GetRunState)
-	rest.WaitAndAssertBackgroundManagerExpiredHeartbeat(t, rt.GetDatabase().ResyncManager)
-
-	_, err = rt.Bucket().Get(base.RolePrefix+"role1", &body)
-	assert.NoError(t, err)
-	role1SeqAfter := body["sequence"].(float64)
-
-	_, err = rt.Bucket().Get(base.UserPrefix+"user1", &body)
-	assert.NoError(t, err)
-	user1SeqAfter := body["sequence"].(float64)
-
-	assert.True(t, role1SeqAfter > role1SeqBefore)
-	assert.True(t, user1SeqAfter > user1SeqBefore)
-
-	for i := 0; i < 10; i++ {
-		docID := fmt.Sprintf("doc%d", i)
-
-		doc, err := rt.GetDatabase().GetSingleDatabaseCollection().GetDocument(base.TestCtx(t), docID, db.DocUnmarshalAll)
-		assert.NoError(t, err)
-
-		assert.True(t, float64(doc.Sequence) > docSeqArr[i])
-	}
-
-	response = rt.SendAdminRequest("GET", "/db/_resync", "")
-	rest.RequireStatus(t, response, http.StatusOK)
-	var resyncStatus db.ResyncManagerResponse
-	err = base.JSONUnmarshal(response.BodyBytes(), &resyncStatus)
-	assert.NoError(t, err)
-	assert.Equal(t, 12, resyncStatus.DocsChanged)
-	assert.Equal(t, 12, resyncStatus.DocsProcessed)
-
-	response = rt.SendAdminRequest("POST", "/db/_online", "")
-	rest.RequireStatus(t, response, http.StatusOK)
-
-	err = rt.WaitForCondition(func() bool {
-		state := atomic.LoadUint32(&rt.GetDatabase().State)
-		return state == db.DBOnline
-	})
-	assert.NoError(t, err)
-
-	// Data is wiped from walrus when brought back online
-	request, _ = http.NewRequest("GET", "/db/_changes?since="+changesResp.LastSeq, nil)
-	request.SetBasicAuth("user1", "letmein")
-	response = rt.Send(request)
-	rest.RequireStatus(t, response, http.StatusOK)
-	err = json.Unmarshal(response.BodyBytes(), &changesResp)
-	assert.Len(t, changesResp.Results, 3)
-	assert.True(t, changesRespContains(changesResp, "userdoc"))
-	assert.True(t, changesRespContains(changesResp, "userdoc2"))
 }
 
 // Single threaded bring DB online
@@ -3708,65 +3181,6 @@ func TestTombstoneCompactionPurgeInterval(t *testing.T) {
 			assert.EqualValues(t, test.expectedPurgeIntervalAfterCompact, dbc.PurgeInterval)
 		})
 	}
-}
-
-// CBG-2150: Tests that resync status is cluster aware
-func TestResyncPersistence(t *testing.T) {
-	if base.UnitTestUrlIsWalrus() {
-		t.Skip("This test only works against Couchbase Server")
-	}
-
-	tb := base.GetTestBucket(t)
-	noCloseTB := tb.NoCloseClone()
-
-	rt1 := rest.NewRestTester(t, &rest.RestTesterConfig{
-		CustomTestBucket: noCloseTB,
-	})
-
-	rt2 := rest.NewRestTester(t, &rest.RestTesterConfig{
-		CustomTestBucket: tb,
-	})
-
-	defer rt2.Close()
-	defer rt1.Close()
-
-	// Create a document to process through resync
-	rt1.CreateDoc(t, "doc1")
-
-	// Start resync
-	resp := rt1.SendAdminRequest("POST", "/db/_offline", "")
-	rest.RequireStatus(t, resp, http.StatusOK)
-
-	rest.WaitAndAssertCondition(t, func() bool {
-		state := atomic.LoadUint32(&rt1.GetDatabase().State)
-		return state == db.DBOffline
-	})
-
-	resp = rt1.SendAdminRequest("POST", "/db/_resync?action=start", "")
-	rest.RequireStatus(t, resp, http.StatusOK)
-
-	// Wait for resync to complete
-	var resyncManagerStatus db.ResyncManagerResponse
-	err := rt1.WaitForCondition(func() bool {
-		resp = rt1.SendAdminRequest("GET", "/db/_resync", "")
-		err := json.Unmarshal(resp.BodyBytes(), &resyncManagerStatus)
-		assert.NoError(t, err)
-
-		if resyncManagerStatus.State == db.BackgroundProcessStateCompleted {
-			return true
-		} else {
-			t.Logf("resyncManagerStatus.State != %v: %v", db.BackgroundProcessStateCompleted, resyncManagerStatus.State)
-			return false
-		}
-	})
-	require.NoError(t, err)
-
-	// Check statuses match
-	resp2 := rt2.SendAdminRequest("GET", "/db/_resync", "")
-	rest.RequireStatus(t, resp, http.StatusOK)
-	fmt.Printf("RT1 Resync Status: %s\n", resp.BodyBytes())
-	fmt.Printf("RT2 Resync Status: %s\n", resp2.BodyBytes())
-	assert.Equal(t, resp.BodyBytes(), resp2.BodyBytes())
 }
 
 // Make sure per DB credentials override per bucket credentials
