+++ rest/api_test.go
@@ -1623,278 +1623,6 @@ func TestAccessOnTombstone(t *testing.T) {
 
 }
 
-// TestSyncFnBodyProperties puts a document into channels based on which properties are present on the document.
-// This can be used to introspect what properties ended up in the body passed to the sync function.
-func TestSyncFnBodyProperties(t *testing.T) {
-
-	base.SetUpTestLogging(t, base.LevelInfo, base.KeyHTTP, base.KeyJavascript)
-
-	const (
-		testDocID   = "testdoc"
-		testdataKey = "testdata"
-	)
-
-	// All of these properties must EXACTLY match the top-level properties seen in the sync function body.
-	// Properties not present in this list, but present in the sync function body will be caught.
-	expectedProperties := []string{
-		testdataKey,
-		db.BodyId,
-		db.BodyRev,
-	}
-
-	// This sync function routes into channels based on top-level properties contained in doc
-	syncFn := `function(doc) {
-		console.log("full doc: "+JSON.stringify(doc));
-		for (var p in doc) {
-			console.log("doc property: "+p);
-			channel(p);
-		}
-	}`
-
-	rtConfig := RestTesterConfig{SyncFn: syncFn, DatabaseConfig: &DatabaseConfig{DbConfig: DbConfig{JavascriptTimeoutSecs: base.Uint32Ptr(0)}}}
-	rt := NewRestTester(t, &rtConfig)
-	defer rt.Close()
-
-	response := rt.SendAdminRequest("PUT", "/db/"+testDocID, `{"`+testdataKey+`":true}`)
-	RequireStatus(t, response, 201)
-
-	syncData, err := rt.GetDatabase().GetSingleDatabaseCollection().GetDocSyncData(base.TestCtx(t), testDocID)
-	assert.NoError(t, err)
-
-	actualProperties := syncData.Channels.KeySet()
-	assert.ElementsMatchf(t, expectedProperties, actualProperties, "Expected sync fn body %q to match expectedProperties: %q", actualProperties, expectedProperties)
-}
-
-// TestSyncFnBodyPropertiesTombstone puts a document into channels based on which properties are present on the document.
-// It creates a doc, and then tombstones it to see what properties are present in the body of the tombstone.
-func TestSyncFnBodyPropertiesTombstone(t *testing.T) {
-
-	base.SetUpTestLogging(t, base.LevelInfo, base.KeyHTTP, base.KeyJavascript)
-
-	const (
-		testDocID   = "testdoc"
-		testdataKey = "testdata"
-	)
-
-	// All of these properties must be present in the sync function body for a tombstone
-	expectedProperties := []string{
-		testdataKey,
-		db.BodyId,
-		db.BodyRev,
-		db.BodyDeleted,
-	}
-
-	// This sync function routes into channels based on top-level properties contained in doc
-	syncFn := `function(doc) {
-		console.log("full doc: "+JSON.stringify(doc));
-		for (var p in doc) {
-			console.log("doc property: "+p);
-			channel(p);
-		}
-	}`
-
-	rtConfig := RestTesterConfig{SyncFn: syncFn}
-	rt := NewRestTester(t, &rtConfig)
-	defer rt.Close()
-
-	response := rt.SendAdminRequest("PUT", "/db/"+testDocID, `{"`+testdataKey+`":true}`)
-	RequireStatus(t, response, 201)
-	var body db.Body
-	assert.NoError(t, base.JSONUnmarshal(response.Body.Bytes(), &body))
-	assert.Equal(t, true, body["ok"])
-	revID := body["rev"].(string)
-
-	response = rt.SendAdminRequest("DELETE", "/db/"+testDocID+"?rev="+revID, `{}`)
-	RequireStatus(t, response, 200)
-
-	syncData, err := rt.GetDatabase().GetSingleDatabaseCollection().GetDocSyncData(base.TestCtx(t), testDocID)
-	assert.NoError(t, err)
-
-	actualProperties := syncData.Channels.KeySet()
-	assert.ElementsMatchf(t, expectedProperties, actualProperties, "Expected sync fn body %q to match expectedProperties: %q", actualProperties, expectedProperties)
-}
-
-// TestSyncFnOldDocBodyProperties puts a document into channels based on which properties are present in the 'oldDoc' body.
-// It creates a doc, and updates it to inspect what properties are present on the oldDoc body.
-func TestSyncFnOldDocBodyProperties(t *testing.T) {
-
-	base.SetUpTestLogging(t, base.LevelInfo, base.KeyHTTP, base.KeyJavascript)
-
-	const (
-		testDocID   = "testdoc"
-		testdataKey = "testdata"
-	)
-
-	// All of these properties must be present in the sync function oldDoc body for a regular PUT containing testdataKey
-	expectedProperties := []string{
-		testdataKey,
-		db.BodyId,
-	}
-
-	// This sync function routes into channels based on top-level properties contained in oldDoc
-	syncFn := `function(doc, oldDoc) {
-		console.log("full doc: "+JSON.stringify(doc));
-		console.log("full oldDoc: "+JSON.stringify(oldDoc));
-		for (var p in oldDoc) {
-			console.log("oldDoc property: "+p);
-			channel(p);
-		}
-	}`
-
-	rtConfig := RestTesterConfig{SyncFn: syncFn}
-	rt := NewRestTester(t, &rtConfig)
-	defer rt.Close()
-
-	response := rt.SendAdminRequest("PUT", "/db/"+testDocID, `{"`+testdataKey+`":true}`)
-	RequireStatus(t, response, 201)
-	var body db.Body
-	assert.NoError(t, base.JSONUnmarshal(response.Body.Bytes(), &body))
-	assert.Equal(t, true, body["ok"])
-	revID := body["rev"].(string)
-
-	response = rt.SendAdminRequest("PUT", "/db/"+testDocID+"?rev="+revID, `{"`+testdataKey+`":true,"update":2}`)
-	RequireStatus(t, response, 201)
-
-	syncData, err := rt.GetDatabase().GetSingleDatabaseCollection().GetDocSyncData(base.TestCtx(t), testDocID)
-	assert.NoError(t, err)
-
-	actualProperties := syncData.Channels.KeySet()
-	assert.ElementsMatchf(t, expectedProperties, actualProperties, "Expected sync fn oldDoc body %q to match expectedProperties: %q", actualProperties, expectedProperties)
-}
-
-// TestSyncFnOldDocBodyPropertiesTombstoneResurrect puts a document into channels based on which properties are present in the 'oldDoc' body.
-// It creates a doc, tombstones it, and then resurrects it to inspect oldDoc properties on the tombstone.
-func TestSyncFnOldDocBodyPropertiesTombstoneResurrect(t *testing.T) {
-
-	base.SetUpTestLogging(t, base.LevelInfo, base.KeyHTTP, base.KeyJavascript)
-
-	const (
-		testDocID   = "testdoc"
-		testdataKey = "testdata"
-	)
-
-	// All of these properties must be present in the sync function body for a regular PUT containing testdataKey
-	expectedProperties := []string{
-		testdataKey,
-		db.BodyId,
-		db.BodyDeleted,
-	}
-
-	// This sync function routes into channels based on top-level properties contained in oldDoc
-	syncFn := `function(doc, oldDoc) {
-		console.log("full doc: "+JSON.stringify(doc));
-		console.log("full oldDoc: "+JSON.stringify(oldDoc));
-		for (var p in oldDoc) {
-			console.log("oldDoc property: "+p);
-			channel(p);
-		}
-	}`
-
-	rtConfig := RestTesterConfig{SyncFn: syncFn}
-	rt := NewRestTester(t, &rtConfig)
-	defer rt.Close()
-
-	response := rt.SendAdminRequest("PUT", "/db/"+testDocID, `{"`+testdataKey+`":true}`)
-	RequireStatus(t, response, 201)
-	var body db.Body
-	assert.NoError(t, base.JSONUnmarshal(response.Body.Bytes(), &body))
-	assert.Equal(t, true, body["ok"])
-	revID := body["rev"].(string)
-
-	response = rt.SendAdminRequest("DELETE", "/db/"+testDocID+"?rev="+revID, `{}`)
-	RequireStatus(t, response, 200)
-	body = nil
-	assert.NoError(t, base.JSONUnmarshal(response.Body.Bytes(), &body))
-	assert.Equal(t, true, body["ok"])
-	revID = body["rev"].(string)
-
-	response = rt.SendAdminRequest("PUT", "/db/"+testDocID+"?rev="+revID, `{"`+testdataKey+`":true}`)
-	RequireStatus(t, response, 201)
-
-	syncData, err := rt.GetDatabase().GetSingleDatabaseCollection().GetDocSyncData(base.TestCtx(t), testDocID)
-	assert.NoError(t, err)
-
-	actualProperties := syncData.Channels.KeySet()
-	assert.ElementsMatchf(t, expectedProperties, actualProperties, "Expected sync fn oldDoc body %q to match expectedProperties: %q", actualProperties, expectedProperties)
-}
-
-// TestSyncFnDocBodyPropertiesSwitchActiveTombstone creates a branched revtree, where the first tombstone created becomes active again after the shorter b branch is tombstoned.
-// The test makes sure that in this scenario, the "doc" body of the sync function when switching from (T) 3-b to (T) 4-a contains a _deleted property (stamped by getAvailable1xRev)
-//
-//	1-a
-//	├── 2-a
-//	│   └── 3-a
-//	│       └──────── (T) 4-a
-//	└──────────── 2-b
-//	              └────────────── (T) 3-b
-func TestSyncFnDocBodyPropertiesSwitchActiveTombstone(t *testing.T) {
-
-	base.SetUpTestLogging(t, base.LevelInfo, base.KeyHTTP, base.KeyJavascript)
-
-	const (
-		testDocID   = "testdoc"
-		testdataKey = "testdata"
-	)
-
-	// This sync function logs a warning for each revision pushed through the sync function, and an error when it sees _deleted inside doc, when oldDoc contains syncOldDocBodyCheck=true
-	//
-	// These are then asserted by looking at the expvar stats for warn and error counts.
-	// We can't rely on channels to get information out of the sync function environment, because we'd need an active doc, which this test does not allow for.
-	syncFn := `function(doc, oldDoc) {
-		console.log("full doc: "+JSON.stringify(doc));
-		console.log("full oldDoc: "+JSON.stringify(oldDoc));
-
-		if (doc.testdata == 1 || (oldDoc != null && !oldDoc.syncOldDocBodyCheck)) {
-			console.log("skipping oldDoc property checks for this rev")
-			return
-		}
-
-		if (doc != null && doc._deleted) {
-			console.error("doc contained _deleted")
-		}
-	}`
-
-	rtConfig := RestTesterConfig{SyncFn: syncFn}
-	rt := NewRestTester(t, &rtConfig)
-	defer rt.Close()
-
-	// rev 1-a
-	resp := rt.SendAdminRequest("PUT", "/db/"+testDocID, `{"`+testdataKey+`":1}`)
-	RequireStatus(t, resp, 201)
-	rev1ID := RespRevID(t, resp)
-
-	// rev 2-a
-	resp = rt.SendAdminRequest("PUT", "/db/"+testDocID+"?rev="+rev1ID, `{"`+testdataKey+`":2}`)
-	RequireStatus(t, resp, 201)
-	rev2aID := RespRevID(t, resp)
-
-	// rev 3-a
-	resp = rt.SendAdminRequest("PUT", "/db/"+testDocID+"?rev="+rev2aID, `{"`+testdataKey+`":3,"syncOldDocBodyCheck":true}`)
-	RequireStatus(t, resp, 201)
-	rev3aID := RespRevID(t, resp)
-
-	// rev 2-b
-	_, rev1Hash := db.ParseRevID(rev1ID)
-	resp = rt.SendAdminRequest("PUT", "/db/"+testDocID+"?new_edits=false", `{"`+db.BodyRevisions+`":{"start":2,"ids":["b", "`+rev1Hash+`"]}}`)
-	RequireStatus(t, resp, 201)
-	rev2bID := RespRevID(t, resp)
-
-	// tombstone at 4-a
-	resp = rt.SendAdminRequest("DELETE", "/db/"+testDocID+"?rev="+rev3aID, `{}`)
-	RequireStatus(t, resp, 200)
-
-	numErrorsBefore, err := strconv.Atoi(base.SyncGatewayStats.GlobalStats.ResourceUtilizationStats().ErrorCount.String())
-	assert.NoError(t, err)
-	// tombstone at 3-b
-	resp = rt.SendAdminRequest("DELETE", "/db/"+testDocID+"?rev="+rev2bID, `{}`)
-	RequireStatus(t, resp, 200)
-
-	numErrorsAfter, err := strconv.Atoi(base.SyncGatewayStats.GlobalStats.ResourceUtilizationStats().ErrorCount.String())
-	assert.NoError(t, err)
-
-	assert.Equal(t, 1, numErrorsAfter-numErrorsBefore, "expecting to see only only 1 error logged")
-}
-
 func TestAllDocsChannelsAfterChannelMove(t *testing.T) {
 
 	type allDocsRow struct {
@@ -2726,36 +2454,6 @@ func TestDocIDFilterResurrection(t *testing.T) {
 	var changesResponse = make(map[string]interface{})
 	require.NoError(t, base.JSONUnmarshal(response.Body.Bytes(), &changesResponse))
 	assert.NotContains(t, changesResponse["results"].([]interface{})[1], "deleted")
-}
-
-func TestSyncFunctionErrorLogging(t *testing.T) {
-
-	base.SetUpTestLogging(t, base.LevelInfo, base.KeyHTTP, base.KeyJavascript)
-
-	rtConfig := RestTesterConfig{SyncFn: `
-		function(doc) {
-			console.error("Error");
-			console.log("Log");
-			channel(doc.channel);
-		}`}
-
-	rt := NewRestTester(t, &rtConfig)
-
-	defer rt.Close()
-
-	// Wait for the DB to be ready before attempting to get initial error count
-	assert.NoError(t, rt.WaitForDBOnline())
-
-	numErrors, err := strconv.Atoi(base.SyncGatewayStats.GlobalStats.ResourceUtilizationStats().ErrorCount.String())
-	assert.NoError(t, err)
-
-	response := rt.SendAdminRequest("PUT", "/db/doc1", `{"foo": "bar"}`)
-	assert.Equal(t, http.StatusCreated, response.Code)
-
-	numErrorsAfter, err := strconv.Atoi(base.SyncGatewayStats.GlobalStats.ResourceUtilizationStats().ErrorCount.String())
-	assert.NoError(t, err)
-
-	assert.Equal(t, numErrors+1, numErrorsAfter)
 }
 
 func TestNumAccessErrors(t *testing.T) {
@@ -3962,24 +3660,6 @@ func TestForceAPIForbiddenErrors(t *testing.T) {
 			assertRespStatus(resp, http.StatusForbidden)
 		})
 	}
-}
-
-func TestSyncFnTimeout(t *testing.T) {
-	syncFn := `function(doc) { while(true) {} }`
-
-	rtConfig := RestTesterConfig{SyncFn: syncFn, DatabaseConfig: &DatabaseConfig{DbConfig: DbConfig{JavascriptTimeoutSecs: base.Uint32Ptr(1)}}}
-	rt := NewRestTester(t, &rtConfig)
-	defer rt.Close()
-
-	syncFnFinishedWG := sync.WaitGroup{}
-	syncFnFinishedWG.Add(1)
-	go func() {
-		response := rt.SendAdminRequest("PUT", "/db/doc", `{"foo": "bar"}`)
-		assertHTTPErrorReason(t, response, 500, "JS sync function timed out")
-		syncFnFinishedWG.Done()
-	}()
-	timeoutErr := WaitWithTimeout(&syncFnFinishedWG, time.Second*15)
-	assert.NoError(t, timeoutErr)
 }
 
 func assertHTTPErrorReason(t testing.TB, response *TestResponse, expectedStatus int, expectedReason string) {
